+++ src/main/java/me/villagerunknown/immortalvillagers/Immortalvillagers.java
@@ -1,6 +1,7 @@
 package me.villagerunknown.immortalvillagers;
 
 import me.villagerunknown.immortalvillagers.feature.nitwitEducationFeature;
+import me.villagerunknown.immortalvillagers.feature.resetVillagerTradesFeature;
 import me.villagerunknown.platform.Platform;
 import me.villagerunknown.immortalvillagers.feature.preventDamageToVillagersFeature;
 import me.villagerunknown.platform.PlatformMod;
@@ -34,5 +35,6 @@ public class Immortalvillagers implements ModInitializer {
 		// # Activate Features
 		featureManager.addFeature( "preventDamageToVillagers", preventDamageToVillagersFeature::execute );
 		featureManager.addFeature( "nitwitEducation", nitwitEducationFeature::execute );
+		featureManager.addFeature( "resetVillagerTrades", resetVillagerTradesFeature::execute );
 	}
 }